Added basic routes for user actions
Added basic routes in routes/users.js for user registration,
authentication and retrieval of some profile info.
Established JWT strategy in passport.js and had bcryptjs handle hashing
of password before storing in database.
Made package versions strict in package.json - may have to revert that
since it was not the cause of non-connectivity with db
Changes to be committed:
	modified:   chedder.js
	modified:   config/database.js
	new file:   config/passport.js
	modified:   package.json
	modified:   routes/users.js